+++ Model/Carrier/ShipmentRequestBuilder.php
@@ -10,9 +10,6 @@ declare(strict_types=1);
 
 namespace GLSCroatia\Shipping\Model\Carrier;
 
-use Magento\Framework\DataObject;
-use Magento\Sales\Model\Order\Shipment;
-
 class ShipmentRequestBuilder
 {
     /**
@@ -26,24 +23,9 @@ class ShipmentRequestBuilder
     protected \GLSCroatia\Shipping\Helper\Data $dataHelper;
 
     /**
-     * @var \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\AddressSwitcher
+     * @var \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\Service
      */
-    protected \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\AddressSwitcher $addressSwitcher;
-
-    /**
-     * @var \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\CashOnDelivery
-     */
-    protected \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\CashOnDelivery $cashOnDelivery;
-
-    /**
-     * @var \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\ExpressDelivery
-     */
-    protected \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\ExpressDelivery $expressDelivery;
-
-    /**
-     * @var \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\Insurance
-     */
-    protected \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\Insurance $insurance;
+    protected \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\Service $service;
 
     /**
      * @var \GLSCroatia\Shipping\ViewModel\ParcelShopDelivery
@@ -53,27 +35,18 @@ class ShipmentRequestBuilder
     /**
      * @param \GLSCroatia\Shipping\Model\Config $config
      * @param \GLSCroatia\Shipping\Helper\Data $dataHelper
-     * @param \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\AddressSwitcher $addressSwitcher
-     * @param \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\CashOnDelivery $cashOnDelivery
-     * @param \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\ExpressDelivery $expressDelivery
-     * @param \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\Insurance $insurance
+     * @param \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\Service $service
      * @param \GLSCroatia\Shipping\ViewModel\ParcelShopDelivery $parcelShopDelivery
      */
     public function __construct(
         \GLSCroatia\Shipping\Model\Config $config,
         \GLSCroatia\Shipping\Helper\Data $dataHelper,
-        \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\AddressSwitcher $addressSwitcher,
-        \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\CashOnDelivery $cashOnDelivery,
-        \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\ExpressDelivery $expressDelivery,
-        \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\Insurance $insurance,
+        \GLSCroatia\Shipping\Model\Carrier\ShipmentRequest\Service $service,
         \GLSCroatia\Shipping\ViewModel\ParcelShopDelivery $parcelShopDelivery
     ) {
         $this->config = $config;
         $this->dataHelper = $dataHelper;
-        $this->addressSwitcher = $addressSwitcher;
-        $this->cashOnDelivery = $cashOnDelivery;
-        $this->expressDelivery = $expressDelivery;
-        $this->insurance = $insurance;
+        $this->service = $service;
         $this->parcelShopDelivery = $parcelShopDelivery;
     }
 
@@ -84,7 +57,7 @@ class ShipmentRequestBuilder
      * @return array
      * @throws \Magento\Framework\Exception\LocalizedException
      */
-    public function getParams(DataObject $request): array
+    public function getParams(\Magento\Framework\DataObject $request): array
     {
         $storeId = $request->getStoreId();
 
@@ -92,13 +65,8 @@ class ShipmentRequestBuilder
             throw new \Magento\Framework\Exception\LocalizedException(__('GLS Client ID is not configured.'));
         }
 
-        // use GLS address if available
-        if ($addressId = $this->config->getAddressId($storeId)) {
-            $request = $this->addressSwitcher->switchShipperAddress((int)$addressId, $request);
-        }
-
         $shipment = $request->getOrderShipment();
-        $order = $shipment->getOrder();
+        $requestData = $request->getData('gls') ?: [];
 
 //        $currentTimestamp = round(microtime(true) * 1000); // milliseconds
 
@@ -146,16 +114,17 @@ class ShipmentRequestBuilder
         $serviceList = [];
 
         // Cash on Delivery Service
-        if ($this->cashOnDelivery->isAllowed($shipment)) {
-            $parcel['CODAmount'] = $this->cashOnDelivery->calculateAmount($shipment);
-            $parcel['CODReference'] = $this->cashOnDelivery->generateReference($shipment);
+        if ($this->service->isCashOnDeliveryAllowed($request)) {
+            $codReference = $requestData['cod_reference'] ?? $this->service->getCashOnDeliveryReference($request);
+
+            $parcel['CODAmount'] = $this->service->getCashOnDeliveryAmount($request);
+            $parcel['CODReference'] = $codReference;
             $serviceList[] = ['Code' => 'COD'];
         }
 
         // Parcel Shop Delivery Service
-        $isShopDeliveryService = $this->isShopDeliveryService((string)$request->getShippingMethod());
-        if ($isShopDeliveryService) {
-            $deliveryPointData = $this->parcelShopDelivery->getParcelShopDeliveryPointData($order);
+        if ($this->isShopDeliveryService((string)$request->getShippingMethod())) {
+            $deliveryPointData = $this->parcelShopDelivery->getParcelShopDeliveryPointData($shipment->getOrder());
 
             $serviceList[] = [
                 'Code' => 'PSD',
@@ -164,32 +133,23 @@ class ShipmentRequestBuilder
         }
 
         // Guaranteed 24h Service
-        if ($this->config->isEnabledGuaranteed24hService($storeId)
-            && $request->getRecipientAddressCountryCode() !== 'RS'
-        ) {
+        $isEnabledGuaranteed24h = $requestData['guaranteed_24h'] ?? $this->config->isEnabledGuaranteed24hService($storeId); // phpcs:ignore
+        if ((bool)$isEnabledGuaranteed24h && $this->service->isGuaranteed24hAllowed($request)) {
             $serviceList[] = ['Code' => '24H'];
         }
 
         // Express Delivery Service
-        $expressDeliverCode = $this->config->getExpressDeliveryServiceCode($storeId);
-        $isExpressDeliveryAllowed = !$isShopDeliveryService
-            && $expressDeliverCode
-            && $this->expressDelivery->isAllowed(
-                $expressDeliverCode,
-                (string)$request->getShipperAddressCountryCode(),
-                (string)$request->getRecipientAddressCountryCode(),
-                (string)$request->getRecipientAddressPostalCode()
-            );
-        if ($isExpressDeliveryAllowed) {
+        $expressDeliverCode = $requestData['express_delivery'] ?? $this->config->getExpressDeliveryServiceCode($storeId); // phpcs:ignore
+        $isExpressDeliveryAllowed = $this->service->isExpressDeliveryAllowed($request, (string)$expressDeliverCode);
+        if ($expressDeliverCode && $isExpressDeliveryAllowed) {
             $serviceList[] = [
                 'Code' => $expressDeliverCode
             ];
         }
 
         // Contact Service
-        if (!$isShopDeliveryService
-            && $this->config->isEnabledContactService($storeId)
-        ) {
+        $isEnabledContact = $requestData['cs1'] ?? $this->config->isEnabledContactService($storeId);
+        if ((bool)$isEnabledContact && $this->service->isContactAllowed($request)) {
             $serviceList[] = [
                 'Code' => 'CS1',
                 'CS1Parameter' => ['Value' => $recipientPhoneNumber]
@@ -197,10 +157,12 @@ class ShipmentRequestBuilder
         }
 
         // Flexible Delivery Service
-        if (!$isShopDeliveryService
-            && !$isExpressDeliveryAllowed
-            && $this->config->isEnabledFlexibleDeliveryService($storeId)
-        ) {
+        $isEnabledFlexibleDelivery = $requestData['fds'] ?? $this->config->isEnabledFlexibleDeliveryService($storeId);
+        $isFlexibleDeliveryAllowed = $this->service->isFlexibleDeliveryAllowed(
+            $request,
+            $isExpressDeliveryAllowed
+        );
+        if ((bool)$isEnabledFlexibleDelivery && $isFlexibleDeliveryAllowed) {
             $serviceList[] = [
                 'Code' => 'FDS',
                 'FDSParameter' => ['Value' => (string)$request->getRecipientEmail()]
@@ -208,11 +170,12 @@ class ShipmentRequestBuilder
         }
 
         // Flexible Delivery SMS Service
-        if (!$isShopDeliveryService
-            && !$isExpressDeliveryAllowed
-            && $this->config->isEnabledFlexibleDeliveryService($storeId)
-            && $this->config->isEnabledFlexibleDeliverySmsService($storeId)
-        ) {
+        $isEnabledFlexibleDeliverySms = $requestData['fss'] ?? $this->config->isEnabledFlexibleDeliverySmsService($storeId); // phpcs:ignore
+        $isFlexibleDeliverySmsAllowed = $this->service->isFlexibleDeliverySmsAllowed(
+            $request,
+            $isEnabledFlexibleDelivery && $isFlexibleDeliveryAllowed
+        );
+        if ((bool)$isEnabledFlexibleDeliverySms && $isFlexibleDeliverySmsAllowed) {
             $serviceList[] = [
                 'Code' => 'FSS',
                 'FSSParameter' => ['Value' => $recipientPhoneNumber]
@@ -220,16 +183,18 @@ class ShipmentRequestBuilder
         }
 
         // SMS Service
-        if ($this->config->isEnabledSmsService($storeId)
-            && $sm1Text = $this->config->getSmsServiceText($storeId)
-        ) {
+        $isEnabledSms = $requestData['sm1'] ?? $this->config->isEnabledSmsService($storeId);
+        $sm1Text = $this->config->getSmsServiceText($storeId);
+        if ((bool)$isEnabledSms && $this->service->isSmsAllowed($request, $sm1Text)) {
             $serviceList[] = [
                 'Code' => 'SM1',
                 'SM1Parameter' => ['Value' => "{$recipientPhoneNumber}|$sm1Text"]
             ];
         }
+
         // SMS Pre-advice Service
-        if ($this->config->isEnabledSmsPreAdviceService($storeId)) {
+        $isEnabledSmsPreAdvice = $requestData['sm2'] ?? $this->config->isEnabledSmsPreAdviceService($storeId);
+        if ((bool)$isEnabledSmsPreAdvice && $this->service->isSmsPreAdviceAllowed($request)) {
             $serviceList[] = [
                 'Code' => 'SM2',
                 'SM2Parameter' => ['Value' => $recipientPhoneNumber]
@@ -237,9 +202,8 @@ class ShipmentRequestBuilder
         }
 
         // Addressee Only Service
-        if (!$isShopDeliveryService
-            && $this->config->isEnabledAddresseeOnlyService($storeId)
-        ) {
+        $isEnabledAddresseeOnly = $requestData['aos'] ?? $this->config->isEnabledAddresseeOnlyService($storeId);
+        if ((bool)$isEnabledAddresseeOnly && $this->service->isAddresseeOnlyAllowed($request)) {
             $serviceList[] = [
                 'Code' => 'AOS',
                 'AOSParameter' => ['Value' => (string)$request->getRecipientContactPersonName()]
@@ -247,16 +211,11 @@ class ShipmentRequestBuilder
         }
 
         // Insurance Service
-        $isInsuranceAllowed = $this->config->isEnabledInsuranceService($storeId)
-            && $this->insurance->isAllowed(
-                $shipment,
-                (string)$request->getShipperAddressCountryCode(),
-                (string)$request->getRecipientAddressCountryCode()
-            );
-        if ($isInsuranceAllowed) {
+        $isEnabledInsurance = $requestData['ins'] ?? $this->config->isEnabledInsuranceService($storeId);
+        if ((bool)$isEnabledInsurance && $this->service->isInsuranceAllowed($request)) {
             $serviceList[] = [
                 'Code' => 'INS',
-                'INSParameter' => ['Value' => $this->insurance->calculateValue($shipment)]
+                'INSParameter' => ['Value' => $this->service->calculateInsuranceValue($request)]
             ];
         }
 
@@ -267,7 +226,7 @@ class ShipmentRequestBuilder
         return [
             'ParcelList' => [$parcel],
             'TypeOfPrinter' => $this->config->getPrinterType($storeId),
-            'PrintPosition' => $this->config->getPrintPosition($storeId),
+            'PrintPosition' => $requestData['print_position'] ?? $this->config->getPrintPosition($storeId),
             'ShowPrintDialog' => false
         ];
     }
@@ -308,10 +267,10 @@ class ShipmentRequestBuilder
     /**
      * Generate "ClientReference" value.
      *
-     * @param Shipment $shipment
+     * @param \Magento\Sales\Model\Order\Shipment $shipment
      * @return string
      */
-    protected function generateReference(Shipment $shipment): string
+    protected function generateReference(\Magento\Sales\Model\Order\Shipment $shipment): string
     {
         return str_replace(
             '{increment_id}',